# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio API",
        "description": "Backend API for the portfolio project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{scheme}://{host}",
            "description": "Dynamic server",
            "variables": {
                "scheme": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/api/v1/chatbot/sessions": {
            "get": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "List chat sessions",
                "operationId": "18c9a89224e4d8e4208a713ffbe59d74",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "Create chat session",
                "operationId": "f88272ffcf77af28232b56083cb727b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/chatbot/sessions/{id}": {
            "get": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "Get chat session by id",
                "operationId": "121cf477935e0d2a9ce13f06f8a8283c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The chat session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "Delete chat session",
                "operationId": "13fd7e7b6669583739f0b2e906271356",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The chat session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/chatbot/sessions/{id}/messages": {
            "get": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "Get chat messages for session",
                "operationId": "da1eb431b01fb63a75a2a63073d2f540",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The chat session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "Send message to chatbot",
                "operationId": "0aef930f69591bb02271b527bb444206",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The chat session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "Clear chat session messages",
                "operationId": "da6cac8ab1498f58927a30514156cbc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The chat session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiError": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "ValidationError"
                    },
                    "code": {
                        "type": "string",
                        "example": "VALIDATION_FAILED"
                    },
                    "details": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "OpenApi",
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "ApiMeta": {
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "page_size": {
                        "type": "integer",
                        "example": 15
                    },
                    "total_count": {
                        "type": "integer",
                        "example": 120
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 8
                    },
                    "has_next_page": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_previous_page": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ApiEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/ApiError"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/ApiMeta"
                    },
                    "trace_id": {
                        "type": "string",
                        "example": "c5d5f3fa-5a7f-4c7a-9c7f-0b2a6e3f2f1e"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Send as: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}